refactor: Add Player abstract class for user, dealer
