+++ src/main/java/blackjack/domain/Player.java
@@ -1,0 +1,25 @@
+package blackjack.domain;
+
+abstract class Player {
+
+    String name;
+    OwnedCards ownedCards;
+
+    public Player() {
+        this.ownedCards = new OwnedCards();
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public OwnedCards getOwnedCards() {
+        return ownedCards;
+    }
+
+    public void addCard(Card card) {
+        ownedCards.addCard(card);
+    }
+
+    abstract boolean isPossibleToGetMoreCard();
+}
